requirement.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==2.2.7
django-adminlte-3==0.1.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
idna==2.10
Paginator==0.5.1
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
